Rename `numericBound()` to `checkNumericBound()` and move the function closer to the parsed type
